refactor: move components to separate files
