+++ package.json
@@ -20,9 +20,10 @@
     "@sentry/react": "7.28.0",
     "@sentry/tracing": "7.28.0",
     "@testing-library/react": "13.4.0",
+    "@types/lodash.debounce": "4.0.7",
     "i18next": "22.4.6",
     "immutable": "4.1.0",
-    "lodash": "^4.17.21",
+    "lodash.debounce": "4.0.8",
     "lodash.isobject": "3.0.2",
     "lodash.isstring": "4.0.1",
     "qs": "6.11.0",
