--- src/App.test.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { render, screen } from '@testing-library/react';
-
-import App from './components/App';
-
-test('renders learn react link', () => {
-  render(<App />);
-  // const linkElement = screen.getByText(/learn react/i);
-  // expect(linkElement).toBeInTheDocument();
-});
