+++ src/components/App.tsx
@@ -1,21 +1,16 @@
-import { FC, ReactElement, useEffect } from 'react';
+import { FC, useEffect } from 'react';
 
 import { useLocalContext } from '@graasp/apps-query-client';
-import { Context } from '@graasp/sdk';
 
-import '../App.css';
 import { DEFAULT_CONTEXT_LANGUAGE } from '../config/appSettings';
 import i18n from '../config/i18n';
-import LoadView from './Excalidraw';
-import Loader from './common/Loader';
+import RenderView from './RenderView';
 import { AppDataProvider } from './context/AppDataContext';
 import { AppSettingProvider } from './context/AppSettingContext';
 import { MembersProvider } from './context/MembersContext';
 
 const App: FC = () => {
   const context = useLocalContext();
-  // eslint-disable-next-line no-console
-  console.log(context.toJS());
 
   useEffect(() => {
     // handle a change of language
@@ -25,25 +20,12 @@ const App: FC = () => {
     }
   }, [context]);
 
-  const renderContent = (): ReactElement => {
-    switch (context.get('context')) {
-      case Context.BUILDER:
-        return <LoadView />;
-      case Context.ANALYTICS:
-        return <Loader />;
-      case Context.PLAYER:
-        // eslint-disable-next-line no-console
-        console.log('rendering for player');
-        return <LoadView />;
-      default:
-        return <LoadView />;
-    }
-  };
-
   return (
     <MembersProvider>
       <AppDataProvider>
-        <AppSettingProvider>{renderContent()}</AppSettingProvider>
+        <AppSettingProvider>
+          <RenderView context={context?.context} />
+        </AppSettingProvider>
       </AppDataProvider>
     </MembersProvider>
   );
