+++ src/components/Root.tsx
@@ -79,6 +79,7 @@ const Root: FC = () => {
   const AppWithContextAndToken = withContext(AppWithContext, {
     LoadingComponent: <Loader />,
     useGetLocalContext: hooks.useGetLocalContext,
+    useAutoResize: hooks.useAutoResize,
     onError:
       /* istanbul ignore next */
       () => {
